examples/sample03/requirements.txt
flask

PyYAML
# For config validation
trafaret
